use base class
