+++ components/script_bindings/codegen/codegen.py
@@ -22,6 +22,14 @@ from itertools import groupby
 from re import Match
 from typing import Any, Generic, Optional, TypeGuard, TypeVar, cast
 
+from cgthings.cgbase import (
+    CGGeneric,
+    CGIndenter,
+    CGThing,
+    CGWrapper,
+    lineStartDetector,
+    stripTrailingWhitespace,
+)
 from configuration import (
     Configuration,
     Descriptor,
@@ -243,16 +251,6 @@ def toBindingModuleFileFromDescriptor(desc: Descriptor) -> str:
     else:
         return toBindingModuleFile(desc.name)
 
-
-def stripTrailingWhitespace(text: str) -> str:
-    tail = '\n' if text.endswith('\n') else ''
-    lines = text.splitlines()
-    for i in range(len(lines)):
-        lines[i] = lines[i].rstrip()
-    joined_lines = '\n'.join(lines)
-    return f"{joined_lines}{tail}"
-
-
 def innerContainerType(type: IDLType) -> IDLType:
     assert type.isSequence() or type.isRecord()
     assert isinstance(type, (IDLSequenceType, IDLRecordType, IDLNullableType))
@@ -309,19 +307,6 @@ numericTags = [
     IDLType.Tags.unrestricted_float,
     IDLType.Tags.unrestricted_double
 ]
-
-
-# We'll want to insert the indent at the beginnings of lines, but we
-# don't want to indent empty lines.  So only indent lines that have a
-# non-newline character on them.
-lineStartDetector = re.compile("^(?=[^\n#])", re.MULTILINE)
-
-
-# We'll want to insert the indent at the beginnings of lines, but we
-# don't want to indent empty lines.  So only indent lines that have a
-# non-newline character on them.
-lineStartDetector = re.compile("^(?=[^\n])", re.MULTILINE)
-
 
 def indent(s: str, indentLevel: int = 2) -> str:
     """
@@ -429,20 +414,6 @@ def fill(template: str, **args: str) -> str:
         args[modified_name] = indent(args[name], depth)
 
     return t.substitute(args)
-
-
-class CGThing():
-    """
-    Abstract base class for things that spit out code.
-    """
-    def __init__(self) -> None:
-        pass  # Nothing for now
-
-    @abstractmethod
-    def define(self) -> str:
-        """Produce code for a Rust file."""
-        raise NotImplementedError
-
 
 class CGMethodCall(CGThing):
     """
@@ -2335,51 +2306,6 @@ class ConstDefiner(PropertyDefiner[IDLConst]):
             'ConstantSpec',
             PropertyDefiner.getControllingCondition, specData)
 
-
-class CGIndenter(CGThing):
-    """
-    A class that takes another CGThing and generates code that indents that
-    CGThing by some number of spaces.  The default indent is two spaces.
-    """
-    def __init__(self, child: CGThing, indentLevel: int = 4) -> None:
-        CGThing.__init__(self)
-        self.child = child
-        self.indent = " " * indentLevel
-
-    def define(self) -> str:
-        defn = self.child.define()
-        if defn != "":
-            return re.sub(lineStartDetector, self.indent, defn)
-        else:
-            return defn
-
-
-class CGWrapper(CGThing):
-    """
-    Generic CGThing that wraps other CGThings with pre and post text.
-    """
-    child: CGThing
-    pre: str
-    post: str
-    reindent: bool
-
-    def __init__(self, child: CGThing, pre: str = "", post: str= "", reindent: bool = False) -> None:
-        CGThing.__init__(self)
-        self.child = child
-        self.pre = pre
-        self.post = post
-        self.reindent = reindent
-
-    def define(self) -> str:
-        defn = self.child.define()
-        if self.reindent:
-            # We don't use lineStartDetector because we don't want to
-            # insert whitespace at the beginning of our _first_ line.
-            defn = stripTrailingWhitespace(
-                defn.replace("\n", f"\n{' ' * len(self.pre)}"))
-        return f"{self.pre}{defn}{self.post}"
-
-
 class CGRecord(CGThing):
     """
     CGThing that wraps value CGThing in record with key type equal to keyType parameter
@@ -2845,20 +2771,6 @@ class CGIfElseWrapper(CGList):
             elseBranch = CGWrapper(CGIndenter(ifFalse), pre=" else {\n", post="\n}")
         kids = [CGIfWrapper(condition, ifTrue), elseBranch]
         CGList.__init__(self, kids)
-
-
-class CGGeneric(CGThing):
-    """
-    A class that spits out a fixed string into the codegen.  Can spit out a
-    separate string for the declaration too.
-    """
-    text: str
-    def __init__(self, text: str) -> None:
-        self.text = text
-
-    def define(self) -> str:
-        return self.text
-
 
 class CGCallbackTempRoot(CGGeneric):
     def __init__(self, name: str) -> None:
